// src/connectors/giphy/models/gif.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Analytics } from '../models/analytics';
import { Images } from '../models/images';
import { User } from '../models/user';

/**
 * The standard GIF Object
 */
export interface Gif {
  alt_text?: string;
  analytics?: Analytics;
  bitly_url?: string;
  content_url?: string;
  create_datetime?: string;
  embed_url?: string;
  id?: string;
  images?: Images;
  import_datetime?: string;
  is_low_contrast?: boolean;
  rating?: string;
  slug?: string;
  source?: string;
  source_post_url?: string;
  source_tld?: string;
  title?: string;
  trending_datetime?: string;
  type?: string;
  update_datetime?: string;
  url?: string;
  user?: User;
  username?: string;
}